Use axes more effectively
This package was started before we had the `axes` infrastructure worked
out in Base. It seems time to do some housecleaning to make better use
of it.

Major changes:
- use `axes` in preference to `size` and add a bunch of deprecations
- generalize `tform2deformation` to work with arbitrary axes
- change the field name from `knots` to `nodes` since these are points
of equal-matching rather than discontinuity.
- clean up some instances of old style indexing, remove some generated
functions
- split the source into multiple files
- split the tests into more testsets

This leaves the old tests as-is, so they issue deprecation warnings. The
next commit will upgrade them to the new syntax.